Move parseEscapeSequence` to `termseq`
