+++ misc_test.go
@@ -7,8 +7,6 @@ import (
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/gdamore/tcell/v2"
 )
 
 func TestIsRoot(t *testing.T) {
@@ -489,22 +487,6 @@ func TestOptionToFmtstr(t *testing.T) {
 	for _, test := range tests {
 		if got := optionToFmtstr(test.s); got != test.exp {
 			t.Errorf("at input %q expected %q but got %q", test.s, test.exp, got)
-		}
-	}
-}
-
-func TestParseEscapeSequence(t *testing.T) {
-	tests := []struct {
-		s   string
-		exp tcell.Style
-	}{
-		{"\033[1m", tcell.StyleDefault.Bold(true)},
-		{"\033[1;7;31;42m", tcell.StyleDefault.Bold(true).Reverse(true).Foreground(tcell.ColorMaroon).Background(tcell.ColorGreen)},
-	}
-
-	for _, test := range tests {
-		if got := parseEscapeSequence(test.s); got != test.exp {
-			t.Errorf("at input %q expected '%v' but got '%v'", test.s, test.exp, got)
 		}
 	}
 }
